--- samples/app-localization/csharp/Localization/Bots/MessageExtension.cs
@@ -1,100 +1,0 @@
-﻿using System.Threading;using System.Threading.Tasks;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Schema;
-using Microsoft.Bot.Builder.Teams;
-using Microsoft.Bot.Schema.Teams;
-using Newtonsoft.Json.Linq;
-using System.Linq;
-using System;
-using System.Collections.Generic;
-using Bogus;
-using Microsoft.Extensions.Localization;
-
-namespace Localization
-{
-    public class MessageExtension : TeamsActivityHandler
-    {
-        private readonly IStringLocalizer<MessageExtension> _localizer;
-
-        public MessageExtension(IStringLocalizer<MessageExtension> localizer)
-        {
-            _localizer = localizer;
-        }
-       
-        protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken)
-        {            
-            turnContext.Activity.RemoveRecipientMention();
-            var text = turnContext.Activity.Text.Trim();        
-
-            var replyText = $"You said:"+ _localizer[text];
-            await turnContext.SendActivityAsync(MessageFactory.Text(replyText, replyText), cancellationToken);            
-        }
-
-        protected override Task<MessagingExtensionResponse> OnTeamsMessagingExtensionQueryAsync(ITurnContext<IInvokeActivity> turnContext, MessagingExtensionQuery query, CancellationToken cancellationToken)
-        {
-            var title = "";
-            var titleParam = query.Parameters?.FirstOrDefault(p => p.Name == "cardTitle");
-            if (titleParam != null)
-            {
-                title = titleParam.Value.ToString();
-            }
-
-            if (query == null || query.CommandId != "getRandomText")
-            {
-                // We only process the 'getRandomText' queries with this message extension
-                throw new NotImplementedException($"Invalid CommandId: {query.CommandId}");
-            }
-
-            var attachments = new MessagingExtensionAttachment[5];
-
-            for (int i = 0; i < 5; i++)
-            {
-                attachments[i] = GetAttachment(title);
-            }
-
-            var result = new MessagingExtensionResponse
-            {
-                ComposeExtension = new MessagingExtensionResult
-                {
-                    AttachmentLayout = "list",
-                    Type = "result",
-                    Attachments = attachments.ToList()
-                },
-            };
-            return Task.FromResult(result);
-
-        }
-
-        private static MessagingExtensionAttachment GetAttachment(string title)
-        {
-            var card = new ThumbnailCard
-            {
-                Title = !string.IsNullOrWhiteSpace(title) ? title : new Faker().Lorem.Sentence(),
-                Text = new Faker().Lorem.Paragraph(),
-                Images = new List<CardImage> { new CardImage("http://lorempixel.com/640/480?rand=" + DateTime.Now.Ticks.ToString()) }
-            };
-
-            return card
-                .ToAttachment()
-                .ToMessagingExtensionAttachment();
-        }
-
-        protected override Task<MessagingExtensionResponse> OnTeamsMessagingExtensionSelectItemAsync(ITurnContext<IInvokeActivity> turnContext, JObject query, CancellationToken cancellationToken)
-        {
-
-            return Task.FromResult(new MessagingExtensionResponse
-            {
-                ComposeExtension = new MessagingExtensionResult
-                {
-                    AttachmentLayout = "list",
-                    Type = "result",
-                    Attachments = new MessagingExtensionAttachment[]{
-                        new ThumbnailCard()
-                            .ToAttachment()
-                            .ToMessagingExtensionAttachment()
-                    }
-                },
-            });
-        }
-    }
-}
