--- samples/app-localization/csharp/Localization/Controllers/LocalizeController.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Localization.Controllers
-{
-    public class LocalizeController
-    {
-    }
-}
