+++ samples/app-localization/csharp/Localization/Resources/Bots/LocalizerBot.es-MX.resx
