+++ samples/app-localization/csharp/Localization/Resources/Bots/LocalizerBot.fr-CA.resx
