+++ samples/app-localization/csharp/Localization/Resources/Bots/LocalizerBot.hi-IN.resx
